dev-3106 relocate files
