+++ src/js/components/award/shared/activity/ActivityXAxis.jsx
