+++ src/js/components/award/shared/activity/ActivityYAxis.jsx
