+++ src/js/components/award/shared/activity/ActivityYAxisItem.jsx
